Move app parsing function to view-utilities package
