+++ packages/view-utilities/src/index.ts
@@ -11,3 +11,4 @@ export * from './normalize-query-keys'
 export * from './convert-price'
 export * from './timestamp'
 export * from './find-folded-position'
+export * from './app-user-agent'
